Complete ZHL signatures for class methods (#815)
Adds 2187 class methods on macOS, 1976 on Linux amd64 and 1948 on Win32
that had no signature before, so they can be hooked or called without
adding each manually. Covers Win32 1.6.9, Linux amd64 1.6.12 and 1.6.13,
and macOS, every store build.
This does not cover global declarations and Linux x86 platform.

New capabilities:

- ZHL keys now include a typeid matcher, so several functions can share
  one name and ZHL picks the right one by its parameters. Before this a
  name alone had to be unique, which is why constructor1/2/3 existed
- zhlscan, verify_sigs.py and find_new_zhl_mismap.py check signatures
  and address drift against a binary on any platform

Changes:

- Rename ZHL names onto the names the binary exports. Names published to
  Lua keep their old spelling, renaming them would break existing mods
- Correct declarations that disagree with the binary: return types,
  parameter types, constness, and parameters the binary never passes
- compare_zhl_nm.py and check_hook_signatures.py compare full types
  instead of names, and run in CI

Bugs this fixed:

- LockdownShard::bDone sat on padding on both amd64 headers, so lockdown
  shards were destroyed at random on macOS and Linux (#760)
- VTable_RepairAnimation resolves to the vtable base on macOS, not the
  vptr, crashing when a repair drone was destroyed (#827)
- On Linux and Windows the three-parameter LockdownShard entry binds to
  a GCC clone that never receives superFreeze, so the hook read a
  garbage value
- The LockdownShard constructor never stores superFreeze in the object
  on any platform, macOS included, so superFreeze was always false
- ShipManager::CanFitSubsystem was hooked as taking an int systemId, but
  the function in the game binary takes no arguments, so the hook read a
  garbage value

---------

+++ TriggeredEvents.cpp
@@ -1035,7 +1035,7 @@ void TriggeredEvent::Reset()
     {
         if (def->maxPlayerHull > def->minPlayerHull)
         {
-            triggerPlayerHull = std::max(triggerPlayerHull, def->minPlayerHull + random32()%(def->maxPlayerHull-def->minPlayerHull+1));
+            triggerPlayerHull = std::max(triggerPlayerHull, def->minPlayerHull + (int)(random32()%(def->maxPlayerHull-def->minPlayerHull+1)));
         }
         else
         {
@@ -1049,7 +1049,7 @@ void TriggeredEvent::Reset()
         int enemyHullScaling = G_->GetWorld()->starMap.worldLevel * def->enemyHullScaling;
         if (def->maxEnemyHull > def->minEnemyHull)
         {
-            triggerEnemyHull = std::max(triggerEnemyHull, def->minEnemyHull + random32()%(def->maxEnemyHull-def->minEnemyHull+1) + enemyHullScaling);
+            triggerEnemyHull = std::max(triggerEnemyHull, def->minEnemyHull + (int)(random32()%(def->maxEnemyHull-def->minEnemyHull+1)) + enemyHullScaling);
         }
         else
         {
@@ -1098,7 +1098,7 @@ void TriggeredEvent::Reset()
     {
         if (def->maxPlayerCrew > def->minPlayerCrew)
         {
-            triggerPlayerCrew = std::max(triggerPlayerCrew, def->minPlayerCrew + random32()%(def->maxPlayerCrew-def->minPlayerCrew+1));
+            triggerPlayerCrew = std::max(triggerPlayerCrew, def->minPlayerCrew + (int)(random32()%(def->maxPlayerCrew-def->minPlayerCrew+1)));
         }
         else
         {
@@ -1111,7 +1111,7 @@ void TriggeredEvent::Reset()
     {
         if (def->maxEnemyCrew > def->minEnemyCrew)
         {
-            triggerEnemyCrew = std::max(triggerEnemyCrew, def->minEnemyCrew + random32()%(def->maxEnemyCrew-def->minEnemyCrew+1));
+            triggerEnemyCrew = std::max(triggerEnemyCrew, def->minEnemyCrew + (int)(random32()%(def->maxEnemyCrew-def->minEnemyCrew+1)));
         }
         else
         {
